Clean up any invalid files in XDG_DESKTOP_DIR
Chromium doesn't directly use flextop tools to create files on the
desktop, but it *does* create the application menu shortcuts after the
invalid desktop file is created, meaning we can just nuke it then. If
the user disables application menu creation, the invalid file will be
removed on next app start (requires CHROME_WRAPPER to be set when
calling flextop-init).

Ideally, *fixing* the desktop files would also be possible, but given
the progress on the dynamic launcher portal, the effort is unlikely to
be worthwhile.